move CachedValue out into package, otherwise new Cache instances referencing them never get gc'd
